+++ aps_lookup.csv
@@ -1,0 +1,8 @@
+CASENO,COUNTRY,AGE,ETHGBEUL,ETH11NI,PWTA18
+14,2,46,11,NA,9667.496677
+15,1,66,8,NA,25871.01269
+16,5,29,NA,6,17152.30926
+17,1,82,9,NA,23203.82533
+18,1,57,1,NA,25404.24882
+19,4,65,1,NA,6139.407165
+20,1,58,8,NA,28241.52192
