+++ country_df0.csv
@@ -1,0 +1,2 @@
+Pnull,TEMP,Pnull,TEMP
+22923.3261344375,1,22923.3261344375,1
