+++ country_df2.csv
@@ -1,0 +1,2 @@
+Pnull,P2,P7
+13152.782278833332,3153.2529768749996,11172.550326454166
